la création des entités et des DTOs nécessaires pour gérer et transférer les données dans l'applicatio

+++ src/main/java/com/example/rh/Models/Entities/Absence.java
@@ -2,6 +2,7 @@ package com.example.rh.Models.Entities;
 
 import jakarta.persistence.*;
 import lombok.*;
+import org.springframework.data.annotation.CreatedDate;
 
 import java.time.LocalDate;
 import java.util.Date;
@@ -16,11 +17,18 @@ import java.util.Date;
 public class Absence {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-    private LocalDate  missingdate;
+    private Long id;
+
+
+    private LocalDate missingdate;
+
+
     private LocalDate Returndate;
+
     private Integer missingduration;
+
     private String missingmotif;
+
     @ManyToOne
     @JoinColumn(name = "personnel_cin")
     private Personnel personnel;
